Don't use fancy Powerline characters at all
I want my terminal to look presentable regardless of whether the server I'm
logged in to has Powerline installed, or whether I've installed a patched fancy
font.  There are too many dependencies outside of the dotfiles themselves to
make it worthwhile.

Are there no codepoints in all of standard Unicode that are a suitable
replacement for Powerline's custom characters?